feat: Reorganized the feture file structure
